refactor: create different app modes with a `mode` module
